requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.11
django-cors-headers==4.3.1
python-decouple==3.8
celery==5.3.4
redis==5.0.1
django-celery-results==2.5.1
django-celery-beat==2.5.0
Pillow==10.1.0
ffmpeg-python==0.2.0
gunicorn==21.2.0
